New flow time matrices. What's most important, they are fully connected now. It is needed for the greedy algorithm. I've also reimplemented the way they work, so they are also faster now (instead of interpolating flow duration for new requestes, matrices for all time frames are computed).

+++ Model/CityBikes.DataProcessing/common/time-processing.hpp
@@ -14,12 +14,12 @@ namespace CityBikes::DataProcessing::Common
 			timeFrames(timeFrames)
 		{ }
 
-		int quantize(std::tm time) const
+		size_t quantize(std::tm time) const
 		{
 			return timeFrames * (time.tm_hour * 60 + time.tm_min) / 1440;
 		}
 
-		int quantize(size_t seconds) const
+		size_t quantize(size_t seconds) const
 		{
 			return timeFrames * seconds / 86400;
 		}
